web: Common: Renamed folder group-selector to GroupSelector
