+++ web/ASC.Web.Common/src/components/GroupSelector/GroupSelector.stories.js
