+++ web/ASC.Web.Common/src/components/GroupSelector/index.js
