+++ web/ASC.Web.Common/src/components/index.js
@@ -5,5 +5,5 @@ export { default as StudioLayout } from "./layout";
 export { default as ModuleTile } from "./module-tile";
 export { default as Headline } from "./headline";
 export { default as PeopleSelector } from "./people-selector";
-export { default as GroupSelector } from "./group-selector";
+export { default as GroupSelector } from "./GroupSelector";
 export { default as AdvancedSelector } from './AdvancedSelector'
